Atomic-Game-Engine-v3.0.0
- Fixed TileMap
- Added Constraint2D in Component Menu
- Changed Icon Objects (Create->3D)
- Updated Duktape
- Updated Monaco Editor
- Fixed CollisionBox2D and CollisionCircle2D.
- Fixed UI design Android Build
- Updated lib.es5.d.ts
- Updated nuget.exe
- Updated Atomic.d.ts
- Added offsets for TileMap
- Updated Image, ResourceCache, Decompress
- Added in third path ETCPACK
- Updated pugiXML
- Updated STB

+++ Source/Atomic/Atomic2D/TileMap2D.h
@@ -48,27 +48,23 @@ public:
     void DrawDebugGeometry(DebugRenderer* debug, bool depthTest) override;
 
     /// Set tmx file.
-    /// @property
     void SetTmxFile(TmxFile2D* tmxFile);
     /// Add debug geometry to the debug renderer.
     void DrawDebugGeometry();
 
     /// Return tmx file.
-    /// @property
     TmxFile2D* GetTmxFile() const;
 
     /// Return information.
-    /// @property
     const TileMapInfo2D& GetInfo() const { return info_; }
 
     /// Return number of layers.
-    /// @property
     unsigned GetNumLayers() const { return layers_.Size(); }
 
     /// Return tile map layer at index.
     TileMapLayer2D* GetLayer(unsigned index) const;
     /// Convert tile index to position.
-    Vector2 TileIndexToPosition(int x, int y) const;
+    Vector2 TileIndexToPosition(float x, float y) const;
     /// Convert position to tile index, if out of map return false.
     bool PositionToTileIndex(int& x, int& y, const Vector2& position) const;
